# perf: use `mimalloc`