Before: 
```sh
┌────────┬────────────┐
│ Name   │ Time Spend │
├────────┼────────────┤
│ mdx-js │ 4,656ms    │
├────────┼────────────┤
│ mdx-rs │ 2,838ms    │
└────────┴────────────┘
```

After:
```sh
┌────────┬────────────┐
│ Name   │ Time Spend │
├────────┼────────────┤
│ mdx-js │ 4,564ms    │
├────────┼────────────┤
│ mdx-rs │ 1,887ms    │
└────────┴────────────┘
```